Pin docker api version to 1.14.
This is only a temporary solution, as soon as docker-py version 1.0.1
is released change version to 'auto'.

Currently all versions greater 0.6. are supported. On the server side,
this means all docker versions greater or equal than 1.2.0

It seems docker currently only enforces that the client version is not
greater than the server version.

This reverts commit 6d21f452d222b14b8542fcafff194414bc68c4ec.