force docker to connect to platform via local interface
user-agent always resolves the domain it tries to connect to.
This makes it hard to try out the pixelated platform because
you would need a valid domain for your platform setup.
This patch forces docker to always connect to the local platform.
It will break if docker does not run on the same host as the platform.